Move library to libraries dir
